bridge/requirements.txt
pywinpty>=2.0
